+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Changed
 - [MAJOR] Updated to ethers v6
+- [MINOR] Added model to abstract away ethers types
 
 ### Removed